requirements.txt
google-generativeai==0.8.3
langchain-google-genai==2.0.1
langchain-community==0.3.2
langchain==0.3.3
pinecone-client==3.0.2
python-dotenv==1.0.1
faiss-cpu==1.9.0
langchain-pinecone==0.2.3
fastapi==0.115.3
uvicorn==0.31.0
sentence-transformers==2.5.1
requests==2.31.0
nest-asyncio==1.5.8 
pymongo==4.10.1